refactor: convert CPU algorithms and loader from double to float

+++ benchmarks/main_benchmark_cpu.cpp
@@ -4,12 +4,12 @@
 #include <iostream>
 
 int main() {
-    std::vector<std::vector<double>> data(10, std::vector<double>(2, 1.0));
+    std::vector<std::vector<float>> data(10, std::vector<float>(2, 1.f));
 
     KMeans km(3, 100);
     km.fit(data);
 
-    PFCM pfcm(3, 100, 2.0, 0.5);
+    PFCM pfcm(3, 100, 2.f, 0.5f);
     pfcm.fit(data);
 
     std::cout << "CPU Benchmark done.\n";
