+++ src/cpu/pfcm_cpu.cpp
@@ -1,10 +1,10 @@
 #include "clustering_core/pfcm.hpp"
 #include <iostream>
 
-PFCM::PFCM(int c, int max_iters, double m, double alpha)
+PFCM::PFCM(int c, int max_iters, float m, float alpha)
     : c_(c), max_iters_(max_iters), m_(m), alpha_(alpha) {}
 
-void PFCM::fit(const std::vector<std::vector<double>>& data) {
+void PFCM::fit(const std::vector<std::vector<float>>& data) {
     std::cout << "PFCM CPU fit() called on " << data.size() << " points\n";
     labels_.resize(data.size(), 0);
 }
