+++ src/cpu/utils.hpp
@@ -4,5 +4,5 @@
 
 void print_vector_size(size_t n);
 
-double euclidean_distance(const std::vector<double>& a,
-                          const std::vector<double>& b);
+float euclidean_distance(const std::vector<float>& a,
+                         const std::vector<float>& b);
